+++ Release_Notes.md
@@ -94,6 +94,9 @@
 - Addressed all ESLint issues
 - Replaced logConsole method with LoggingService.console in the modal.service
 - ObjectFactor.Create can now create an object using a constructor with or with out parameters.
+- DAccounts Added a constructor that accepts connectionString and securityEntitySeqID so we don't have to set the properties in BAccounts.
+- BAccounts changed the constructor to pass the connectionString and securityEntitySeqID to the DAccounts constructor
+- BAccounts now accounts for CentralManagement
 
 ### Known Bugs
 - There is an issue where the refresh token is being revoked when it shouldn't be causing a logout to occure